Add files via upload
This is for review before telling the company doing the work on the layout to move forward! Please use the free PADS viewer and report anything you see that doesnt belong or is wrong. The company understands its not our job to review the work they do, but i feel we should provide some feedback before the final push to the end is started.